[Refactor]: Refactor ActionService.
